# Adding interface for batched integer gemm