--- a/onnxruntime/contrib_ops/cpu/quantization/attention_quant.cc
+++ b/onnxruntime/contrib_ops/cpu/quantization/attention_quant.cc
@@ -190,53 +190,56 @@ Status QAttention<T>::Compute(OpKernelContext* context) const {
     const auto* weights_data = packed_weights_ ? nullptr : static_cast<const uint8_t*>(weights->DataRaw());
     const bool weights_is_signed = packed_weights_ ? weights_is_signed_ : weights->IsDataType<int8_t>();
 
-    const double cost =
-        static_cast<double>(sequence_length) * static_cast<double>(head_size) * static_cast<double>(input_hidden_size);
-    ThreadPool::TryParallelFor(tp, loop_len, cost, [&](std::ptrdiff_t begin, std::ptrdiff_t end) {
-      for (std::ptrdiff_t i = begin; i != end; ++i) {
-        const int batch_index = static_cast<int>((i / 3) / num_heads_);
-        const int head_index = static_cast<int>((i / 3) % num_heads_);
-        const int qkv_index = static_cast<int>(i % 3);
+    MLAS_GEMM_U8X8_SHAPE_PARAMS gemm_shape;
+    gemm_shape.M = sequence_length;
+    gemm_shape.N = head_size;
+    gemm_shape.K = input_hidden_size;
+    gemm_shape.BIsSigned = weights_is_signed;
 
-        int input_offset = batch_index * sequence_length * input_hidden_size;
-        int weights_offset = qkv_index * hidden_size + head_index * head_size;
-        float* qkv_dest = QKV[qkv_index];
-        int qkv_offset = (batch_index * num_heads_ + head_index) * (sequence_length * head_size);
+    std::vector<MLAS_GEMM_U8X8_DATA_PARAMS> gemm_data_vec(loop_len);
+    std::vector<MLAS_QGEMM_SCALE_BIAS_OUTPUT_PROCESSOR> scale_bias_procs;
+    scale_bias_procs.reserve(loop_len);
 
-        //                   original           transposed            iteration
-        // A: input          (BxSxD)            (B.)S x D             S x D
-        // B: weights        (Dx3xNxH)          D  x (3.N.)H          D x H
-        // C: QKV[qkv_index] (3xBxNxSxH)        (3.B.N.)S x H         S x H
+    for (int i = 0; i < loop_len; i++) {
+      const int batch_index = static_cast<int>((i / 3) / num_heads_);
+      const int head_index = static_cast<int>((i / 3) % num_heads_);
+      const int qkv_index = static_cast<int>(i % 3);
 
-        MLAS_QGEMM_SCALE_BIAS_OUTPUT_PROCESSOR scale_bias_processor(qkv_dest + qkv_offset,
-                                                                    head_size,
-                                                                    &dequant_scale,
-                                                                    bias_data + weights_offset);
+      int input_offset = batch_index * sequence_length * input_hidden_size;
+      int weights_offset = qkv_index * hidden_size + head_index * head_size;
+      float* qkv_dest = QKV[qkv_index];
+      int qkv_offset = (batch_index * num_heads_ + head_index) * (sequence_length * head_size);
 
-        MLAS_GEMM_U8X8_PARAMETERS gemm_params;
-        gemm_params.M = sequence_length;
-        gemm_params.N = head_size;
-        gemm_params.K = input_hidden_size;
-        gemm_params.A = input_data + input_offset;
-        gemm_params.lda = input_hidden_size;
-        gemm_params.ZeroPointA = input_zero_point;
-        if (packed_weights_) {
-          const auto* packed_weight =
-              static_cast<const uint8_t*>(packed_weights_.get()) + packed_weights_size_ * (weights_offset / head_size);
-          gemm_params.B = packed_weight;
-          gemm_params.BIsPacked = true;
-        } else {
-          gemm_params.B = weights_data + weights_offset;
-          gemm_params.ldb = 3 * hidden_size;
-        }
-        gemm_params.ZeroPointB = &weight_zero_point;
-        gemm_params.BIsSigned = weights_is_signed;
-        gemm_params.C = reinterpret_cast<int32_t*>(qkv_dest + qkv_offset);
-        gemm_params.ldc = head_size;
-        gemm_params.OutputProcessor = &scale_bias_processor;
-        MlasGemm(&gemm_params, nullptr);
+      //                   original           transposed            iteration
+      // A: input          (BxSxD)            (B.)S x D             S x D
+      // B: weights        (Dx3xNxH)          D  x (3.N.)H          D x H
+      // C: QKV[qkv_index] (3xBxNxSxH)        (3.B.N.)S x H         S x H
+
+      scale_bias_procs.emplace_back(qkv_dest + qkv_offset,
+                                    head_size,
+                                    &dequant_scale,
+                                    bias_data + weights_offset);
+
+      auto& gemm_params = gemm_data_vec[i];
+      gemm_params.A = input_data + input_offset;
+      gemm_params.lda = input_hidden_size;
+      gemm_params.ZeroPointA = input_zero_point;
+      if (packed_weights_) {
+        const auto* packed_weight =
+            static_cast<const uint8_t*>(packed_weights_.get()) + packed_weights_size_ * (weights_offset / head_size);
+        gemm_params.B = packed_weight;
+        gemm_params.BIsPacked = true;
+      } else {
+        gemm_params.B = weights_data + weights_offset;
+        gemm_params.ldb = 3 * hidden_size;
       }
-    });
+      gemm_params.ZeroPointB = &weight_zero_point;
+      gemm_params.C = reinterpret_cast<int32_t*>(qkv_dest + qkv_offset);
+      gemm_params.ldc = head_size;
+      gemm_params.OutputProcessor = &(scale_bias_procs[i]);  
+    }
+
+    MlasGemmBatch(gemm_shape, gemm_data_vec.data(), loop_len, tp);
   }
 
   // Compute the attention score and apply the score to V