… onnxruntime/core/providers/cpu/math/matmul_integer.cc
… onnxruntime/core/providers/cpu/math/quantize_linear_matmul.cc
… onnxruntime/core/providers/cpu/nn/conv_integer.cc
… onnxruntime/core/providers/cpu/nn/qlinearconv.cc
… onnxruntime/core/providers/cpu/rnn/rnn_helpers.cc
… onnxruntime/core/providers/cpu/tensor/dynamicquantizelinear.cc
… onnxruntime/core/providers/cpu/tensor/quantize_linear.cc
… onnxruntime/core/util/qmath.h
… onnxruntime/test/contrib_ops/quantize_lstm_op_test.cc
… onnxruntime/test/mlas/unittest/test_qgemm.h
… onnxruntime/test/mlas/unittest/test_qgemm_fixture.h
… onnxruntime/test/onnx/microbenchmark/quantize.cc
… onnxruntime/test/onnx/microbenchmark/reduceminmax.cc