**Description**: Parallelize MinMax, Quantize and batched quantize GEMM

**Motivation and Context**
Performance problem identified in T5 decoder model (quantized).  DynamicMatMul operator is identified as the culprit. This operator spend time on getting MinMax of a Tensor, quantize a tensor, and perform a batched gemm. All of these can be parallelized.

Currently GEMM is parallelized. However, in batched GEMM, we sequentially call GEMM multiple times. This causes multiple starting and ending of parallel sections, which can be slow sometimes.  We made the following changes:

1. Parallel task partition no longer depends on degree of parallelism, only on shape of the matrices.
2. In a single GEMM, perform 2D partition of the multiplication, along panel lines, to reduce repeated packing.
3. For batched GEMM, all parallel tasks are executed in a single parallel section, reducing the cost of starting threads and waiting for them to finish.

 In GetQuantizeParameters function, MinMax was executed sequentially.  In this change, minmax was parallelized: (all time unit ns, benchmark parameters {array_size, threads}

Baseline | --       | New Implementation | --
--------- | -------: | ------ | --------------:
BM_GetQuantParams/80000/real_time | 6,525 | -- | --
-- | -- | BM_GetQuantParams/80000/2/real_time | 6,696
-- | -- | BM_GetQuantParams/80000/4/real_time | 6,727
-- | -- | BM_GetQuantParams/80000/6/real_time | 6,753
-- | -- | BM_GetQuantParams/80000/8/real_time | 6,716
BM_GetQuantParams/160000/real_time | 13,033 | -- | --
-- | -- | BM_GetQuantParams/160000/2/real_time | 8,281
-- | -- | BM_GetQuantParams/160000/4/real_time | 6,146
-- | -- | BM_GetQuantParams/160000/6/real_time | 6,073
-- | -- | BM_GetQuantParams/160000/8/real_time | 8,457
BM_GetQuantParams/320000/real_time | 34,508 | -- | --
-- | -- | BM_GetQuantParams/320000/2/real_time | 14,693
-- | -- | BM_GetQuantParams/320000/4/real_time | 9,863
-- | -- | BM_GetQuantParams/320000/6/real_time | 9,648
-- | -- | BM_GetQuantParams/320000/8/real_time | 9,439
BM_GetQuantParams/640000/real_time | 97,008 | -- | --
-- | -- | BM_GetQuantParams/640000/2/real_time | 37,335
-- | -- | BM_GetQuantParams/640000/4/real_time | 16,848
-- | -- | BM_GetQuantParams/640000/6/real_time | 14,924
-- | -- | BM_GetQuantParams/640000/8/real_time | 13,268
BM_GetQuantParams/1280000/real_time | 193,904 | -- | --
-- | -- | BM_GetQuantParams/1280000/2/real_time | 99,008
-- | -- | BM_GetQuantParams/1280000/4/real_time | 43,086
-- | -- | BM_GetQuantParams/1280000/6/real_time | 31,400
-- | -- | BM_GetQuantParams/1280000/8/real_time | 23,842



Parallelization of Quantization function:  (all time unit ns, benchmark parameters {size, threads})


Baseline | ns | New Implementation | ns
-- | --: | -- | --:
BM_Quantize/80000/real_time | 14,982 | BM_Quantize/80000/2/real_time | 9,244
  |   | BM_Quantize/80000/4/real_time | 9,642
  |   | BM_Quantize/80000/6/real_time | 7,043
  |   | BM_Quantize/80000/8/real_time | 7,677
BM_Quantize/160000/real_time | 29,869 | BM_Quantize/160000/2/real_time | 16,680
  |   | BM_Quantize/160000/4/real_time | 10,958
  |   | BM_Quantize/160000/6/real_time | 10,217
  |   | BM_Quantize/160000/8/real_time | 10,917
BM_Quantize/320000/real_time | 63,170 | BM_Quantize/320000/2/real_time | 31,690
  |   | BM_Quantize/320000/4/real_time | 18,806
  |   | BM_Quantize/320000/6/real_time | 15,859
  |   | BM_Quantize/320000/8/real_time | 14,871
BM_Quantize/640000/real_time | 126,762 | BM_Quantize/640000/2/real_time | 64,959
  |   | BM_Quantize/640000/4/real_time | 34,477
  |   | BM_Quantize/640000/6/real_time | 26,650
  |   | BM_Quantize/640000/8/real_time | 22,863
BM_Quantize/1280000/real_time | 256,442 | BM_Quantize/1280000/2/real_time | 132,517
  |   | BM_Quantize/1280000/4/real_time | 70,511
  |   | BM_Quantize/1280000/6/real_time | 50,509
  |   | BM_Quantize/1280000/8/real_time | 38,917



Batched QGemm. Existing implementation: multiple multiplication operations are executed sequentially, while each matric multiplication maybe parallelized. In the new implementation, we tried to parallelize the whole batch. Benchmark shows speedup of batched gemm, especially when B is not pre-packed, and when number of threads is high.


  | Baseline | New
-- | --: | --:
QGEMM/PackB/M:512/N:64/K:512/Batch:12/Threads:4 | 266821 | 223233
QGEMM/PackB/M:512/N:512/K:64/Batch:12/Threads:4 | 352279 | 333067
QGEMM/PackB/M:128/N:768/K:2304/Batch:1/Threads:4 | 255532 | 250407
QGEMM/PackB/M:128/N:768/K:2304/Batch:6/Threads:4 | 1717200 | 1763590
QGEMM/PackB/M:128/N:1024/K:4096/Batch:1/Threads:4 | 682238 | 630677
QGEMM/PackB/M:128/N:1024/K:4096/Batch:6/Threads:4 | 4184540 | 4355130
QGEMM/PackB/M:512/N:64/K:512/Batch:12/Threads:8 | 189041 | 169766
QGEMM/PackB/M:512/N:512/K:64/Batch:12/Threads:8 | 239872 | 231030
QGEMM/PackB/M:128/N:768/K:2304/Batch:1/Threads:8 | 195874 | 196617
QGEMM/PackB/M:128/N:768/K:2304/Batch:6/Threads:8 | 1184220 | 998407
QGEMM/PackB/M:128/N:1024/K:4096/Batch:1/Threads:8 | 307591 | 310912
QGEMM/PackB/M:128/N:1024/K:4096/Batch:6/Threads:8 | 2873200 | 2397180
QGEMM/PackB/M:512/N:64/K:512/Batch:12/Threads:16 | 205724 | 113060
QGEMM/PackB/M:512/N:512/K:64/Batch:12/Threads:16 | 228505 | 173621
QGEMM/PackB/M:128/N:768/K:2304/Batch:1/Threads:16 | 98155 | 96369.3
QGEMM/PackB/M:128/N:768/K:2304/Batch:6/Threads:16 | 875479 | 730881
QGEMM/PackB/M:128/N:1024/K:4096/Batch:1/Threads:16 | 223546 | 219434
QGEMM/PackB/M:128/N:1024/K:4096/Batch:6/Threads:16 | 2046790 | 1687330
-- | --: | --:
QGEMM/NoPackB/M:512/N:64/K:512/Batch:12/Threads:4 | 297499 | 246691
QGEMM/NoPackB/M:512/N:512/K:64/Batch:12/Threads:4 | 371799 | 329250
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:1/Threads:4 | 416085 | 430584
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:6/Threads:4 | 3214770 | 2056690
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:1/Threads:4 | 962293 | 959191
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:6/Threads:4 | 7631920 | 4983940
QGEMM/NoPackB/M:512/N:64/K:512/Batch:12/Threads:8 | 240266 | 184689
QGEMM/NoPackB/M:512/N:512/K:64/Batch:12/Threads:8 | 279545 | 237145
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:1/Threads:8 | 188994 | 186943
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:6/Threads:8 | 2341790 | 1155460
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:1/Threads:8 | 435023 | 422203
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:6/Threads:8 | 6108640 | 2771390
QGEMM/NoPackB/M:512/N:64/K:512/Batch:12/Threads:16 | 253414 | 122915
QGEMM/NoPackB/M:512/N:512/K:64/Batch:12/Threads:16 | 279043 | 182560
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:1/Threads:16 | 126859 | 120570
QGEMM/NoPackB/M:128/N:768/K:2304/Batch:6/Threads:16 | 1895960 | 1074540
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:1/Threads:16 | 313716 | 299089
QGEMM/NoPackB/M:128/N:1024/K:4096/Batch:6/Threads:16 | 4511390 | 2323350